requirements.hf.txt
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==14.3.0
bitsandbytes==0.45.5
boto3==1.37.37
botocore==1.37.37
cached_path==1.7.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
ctranslate2==4.6.0
cycler==0.12.1
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
f5-tts==1.1.0
fastapi==0.115.12
faster-whisper==1.1.1
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2024.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.14.0
gradio_client==1.7.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mistral_common==1.5.4
mistral_inference==1.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
omegaconf==2.3.0
onnxruntime==1.21.1
orjson==3.10.16
packaging==25.0
pandas==2.2.3
peft==0.15.2
pillow==11.0.0
platformdirs==4.3.7
pooch==1.8.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypinyin==0.54.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.6
s3transfer==0.11.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spaces==0.35.0
starlette==0.46.2
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.50.3
transformers-stream-generator==0.0.5
trl==0.16.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vocos==0.1.0
wandb==0.19.9
websockets==14.2
win32_setctime==1.2.0
x-transformers==2.2.12
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.20.0
